Phase 9: clear view.cc's and do_print.cc's compiler warnings
The last file with any remaining warnings after the calcalcs.cc/
file_netcdf.cc/ncview.cc/overlay.cc/util.cc sweep. Covers all four
remaining categories at once, since they're all concentrated in
view.cc:

- -Wsign-compare (12 sites): loop indices retyped from int/long to
  size_t where the loop only ever counts forward against a
  size_t-typed bound and the index isn't also used in a signed
  context elsewhere in the same function (view_draw's, do_print's,
  view_check_new_data's, and view_data_edit's plain i/j loop
  counters). Left `long i` alone in plot_XY_sc(), which genuinely
  counts down through 0 with `i >= 0` -- retyping it would either
  infinite-loop or need restructuring, neither of which belongs in a
  warnings-only sweep. The handful of sites where the *bound* (not the
  index) is the mismatched operand (view->variable->n_dims, an int;
  options.blowup_default_size, an int; n_new_frame_times-1, an int)
  got a local static_cast<size_t> at the comparison instead of a wider
  retype, since those fields are shared across many call sites already
  established as `int` in defines.h.
- -Wstringop-truncation (3 sites) + -Wformat-truncation= companion:
  view_report_position()'s xdim_str/ydim_str fixed char[80] buffers
  converted to std::string; view_construct_scalar_coord_str()'s two
  strncat() appends into its char* out-param replaced with equivalent
  snprintf() calls (same computed remaining-length argument, so
  identical effective truncation point), which GCC's stringop-
  truncation analysis doesn't flag the way it flags strncat/strncpy.
  One incidental behavior note: the old strncpy(xdim_str, temp_string,
  80) had no room for a null terminator if temp_string was >=80 bytes
  long (a latent buffer overread on the following %s), so it was never
  possible to preserve that exact behavior in a defined way; the
  std::string conversion truncates at 79 bytes instead, matching the
  same function's other branch (snprintf(..., 79, ...)). Unreachable
  in practice -- these format dimension-value strings, which are never
  remotely that long -- but noted per this branch's inline-fix-with-
  justification rule for cases too entangled to preserve verbatim.
- -Wempty-body (1 site): view_draw()'s empty `else` statement (a
  bare `;` with only a comment) wrapped in braces.
- -Wformat= (1 site): a debug-only printf's `%d` argument was
  `ViewDataStatus`, an enum class; added the missing
  static_cast<int>().

Verified: full rebuild is completely warning-free (0 warnings, down
from 214 at Phase 8), ctest 100% (2/2, incl. the byte-identical Xvfb
ncview_ui_smoke screenshot test).

+++ core/src/do_print.cc
@@ -112,7 +112,7 @@ print_init( void )
 	void
 do_print( void )
 {
-	long	i, j;
+	size_t	i, j;
 	size_t	x_size, y_size, scaled_x_size, scaled_y_size, top_of_image, bot_of_image, 
 		center_x, center_y, left_of_image, right_of_image;
 	char	outfname[1024], tstr[1500];
